chore: auth refactor
