+++ src/Routes.js
@@ -13,7 +13,7 @@ import {
   SupportPage,
   AddProducts,
   EditProducts,
-  PasswordChange,
+  PasswordSucess,
   StoreCreated,
   PasswordReset,
 } from "./pages";
@@ -39,7 +39,7 @@ const MyRoutes = () => {
         <Route exact path="/roles-and-permissions/add-new-role" element={<AddNewRole />} />
         <Route exact path="/profile" element={<ProfilePage />} />
         <Route exact path="/products/edit/:sku" element={<EditProducts />} />
-        <Route exact path="/reset-successful" element={<PasswordChange />} />
+        <Route exact path="/reset-successful" element={<PasswordSucess />} />
         <Route exact path="/store-created" element={<StoreCreated />} />
         <Route exact path="/reset-password" element={<PasswordReset />} />
       </Routes>
