--- src/pages/PasswordChange.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { AuthLayout, ResetSuccessful } from "../components";
-
-const PasswordChange = () => {
-  return (
-    <AuthLayout>
-      <ResetSuccessful />
-    </AuthLayout>
-  );
-};
-
-export default PasswordChange;
