+++ src/pages/index.js
@@ -2,7 +2,6 @@ export { default as OverviewPage } from './App'
 export { default as Orders } from './orders-dashboard'
 export { default as Products } from './products-dashboard'
 export { default as ViewOrdersPage } from './view-orders'
-// export { default as AddProducts } from './AddProducts'
 export { default as AddProducts } from './add-product'
 export { default as RolesAndPermissionsDashboard } from "./roles-and-permissions/roles-and-permissions-dashboard";
 export { default as AddNewRole } from "./roles-and-permissions/add-new-role";
@@ -10,6 +9,6 @@ export {default as DashboardPage} from './dashboard'
 export { default as SupportPage} from './support'
 export { default as ProfilePage } from "./profile";
 export { default as EditProducts } from './edit-products'
-export { default as PasswordChange} from './PasswordChange';
-export { default as StoreCreated} from './StoreCreated'
-export { default as PasswordReset} from './PasswordReset';
+export { default as PasswordSucess} from './password-success';
+export { default as StoreCreated} from './store-create-success'
+export { default as PasswordReset} from './password-reset';
