+++ src/pages/password-success.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { AuthLayout, AuthSuccess } from "../components";
+
+const PasswordSucess = () => {
+  return (
+    <AuthLayout>
+      <AuthSuccess text="Password Change Successful!"/>
+    </AuthLayout>
+  );
+};
+
+export default PasswordSucess;
